steamcompmgr_shared.hpp: make GetGeometry() codegen more compact & efficient (on GCC)
For some reason, I think this function got worse after gamescope incorporated in lua, due to the need to compile w/ -fno-exceptions which was a new flag that had to be added in order for sol.hpp to work right.

that's the only possible explanation I can think of, because, iirc, GetGeometry() hasn't ever been changed since it was introduced, and I distincly remember the function not being as long as it is currently.